requirements.txt
ipython>=8.10
jupyterlab>=3.0
notebook
matplotlib
seaborn

snowflake-connector-python[pandas]>=3.10.0
snowflake-snowpark-python[pandas]>=1.20.0
kedro>=0.19.5
kedro-datasets[snowflake]>=1.7.0

certifi>=2025.8.3
boto3>=1.35.0
botocore>=1.35.0

scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.4.0

